requirements.txt
django
python-dotenv
mysqlclient
django-jazzmin
python-telegram-bot